% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggLD.R
\name{ggLD}
\alias{ggLD}
\title{Visualize correlation with the diagonal on the x-axis}
\usage{
ggLD(data)
}
\arguments{
\item{data}{input correlation matrix}
}
\value{
ggplot2 object
}
\description{
Visualize correlation with the diagonal on the x-axis
}
\examples{
ggLD(data = abs(cor(data.frame(a = rnorm(20), b = rnorm(20), c = rnorm(20)))))

}
